requirements.txt
asgiref==3.3.1
Django==3.1.5
django-cors-headers==3.6.0
django-db-prefix==1.0.5
django-rest-passwordreset==1.1.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
mysqlclient==2.0.3
PyJWT==2.0.1
pytz==2020.5
sqlparse==0.4.1
asgiref==3.3.1
Django==3.1.5
django-cors-headers==3.6.0
django-db-prefix==1.0.5
django-filter==2.4.0
django-rest-passwordreset==1.1.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
mysqlclient==2.0.3
Pillow==8.1.0
PyJWT==2.0.1
pytz==2020.5
sqlparse==0.4.1
